Move unit test to correct package
