+++ sharding-spring/sharding-jdbc-orchestration-spring/sharding-jdbc-orchestration-spring-boot-starter/src/test/java/org/apache/shardingsphere/shardingjdbc/orchestration/spring/boot/util/EmbedTestingServer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingjdbc.spring.boot.util;
+package org.apache.shardingsphere.shardingjdbc.orchestration.spring.boot.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
