+++ sharding-spring/sharding-jdbc-orchestration-spring/sharding-jdbc-orchestration-spring-boot-starter/src/test/resources/META-INF/services/org.apache.shardingsphere.orchestration.reg.api.RegistryCenter
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.shardingjdbc.spring.boot.registry.TestRegistryCenter
+org.apache.shardingsphere.shardingjdbc.orchestration.spring.boot.registry.TestRegistryCenter
